Define logging system
One has to make sure build Json is never called in the
global scope. Otherwise the JsonObject never gets
destroyed an takes up stack space.
The key "type" describes the type of the message and
has to match the keyswords on the receiving end
because they are used to define what the RPi is supposed to do.