Implement python test framework
The inputs are simulated via RPi and relayboard.
The output is only checked via the serial communication
so no checks are performed on the logging code itself.
Also the "feature" of the controllino resetting when serial connection
is lost is used as a way to reset the controllino before each test.
Right now each test sets on input to on and checks weather
the response with the correct type is being captured and
compares the relevant outputs with the expected outputs.
This all depend heavily on the correct names for the log messages
as well as for the pins itself.