Add mapping method to Machine class and make it work
Machine.map takes input and output in various forms (overloaded).
Ich checks wether input matches state of input pins.
If so it sets the state of the output pin to the given output.
Tests have been updated to work with new routine and logging messages.
Right now the sucess of all tests depends on how much time the controllino has to reboot.
This should be changed somehow but does not change the outcome at the moment.