Add communication over Ethernet and handle edgecases.
All communication with the rpi is done over ethernet  now using tcp protocol.
Right now if the server (rpi) restarts the client (controllino)
does not try to reconnect itself. This could be helpfull but is not a must.
Inputs.getChanges now also checks if the changed pinstate
is not the same as a predefiend expected input state.
This way there is only a response if there really is a sensorerror.
Also normal operation is now defiend in init of outputs.
The check  before normal operation is set now has an own method in Machine
which is called in setup() AFTER Machine.begin().